src/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
azure-cosmos==4.7.0
azure-storage-blob==12.22.0
python-dotenv==1.0.1
